+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to Condense are documented in this file. This project follow
 
 ## [0.3.3] - 2026-06-30
 
+### Added
+- New Condense ascii logo to cli
+
 ### Changed
 - Updated `js-yaml` from 5.1.0 to 5.2.0.
 - Updated `prettier` from 3.8.4 to 3.9.4